Break Renderer out into its own file
